move TextDemoModel to shared
